handles principals pagination controller same as users
